+++ tests/lib/Service/OpenProjectAPIServiceTest.php
@@ -742,8 +742,8 @@ class OpenProjectAPIServiceTest extends TestCase {
 		$configMock = null,
 		$tokenProviderMock = null,
 		$db = null,
-		$iURLGenerator = null,
-		$iLogFactory = null
+		$iLogFactory = null,
+		$iURLGenerator = null
 	): OpenProjectAPIService {
 		$onlyMethods[] = 'getBaseUrl';
 		if ($rootMock === null) {
@@ -3558,6 +3558,7 @@ class OpenProjectAPIServiceTest extends TestCase {
 			$configMock,
 			null,
 			null,
+			null,
 			$iULGeneratorMock
 		);
 		$imageURL = 'http://nextcloud/server/index.php/apps/integration_openproject/avatar?userId=3&userName=OpenProject Admin';
@@ -3617,5 +3618,98 @@ class OpenProjectAPIServiceTest extends TestCase {
 		);
 		$resultGetWorkPackageInfo = $service->getWorkPackageInfo('testUser', 123);
 		$this->assertNull($resultGetWorkPackageInfo);
+	}
+
+	public function auditLogDataProvider(): array {
+		return [
+			[
+				// log level less than 1
+				0,
+				\OC::$SERVERROOT . '/data/audit.log',
+				[],
+				true,
+				false
+			],
+			[
+				// wrong path to audit file
+				1,
+				'wrong-path-to-audit-log/audit.log',
+				[],
+				true,
+				false
+			],
+			[
+				// no audit apps in apps section
+				1,
+				'/audit.log',
+				['apps' => ['no_audit_app']],
+				true,
+				false
+			],
+			[
+				// multiple values including apps with no audit apps in apps section
+				1,
+				'/audit.log',
+				['apps' => ['no_audit_app'], 'others' => []],
+				true,
+				false
+			],
+			[
+				// all values configured correctly
+				1,
+				'/audit.log',
+				['apps' => ['admin_audit']],
+				true,
+				true
+			],
+			[
+				// admin_audit app not installed
+				1,
+				'/audit.log',
+				['apps' => ['admin_audit']],
+				false,
+				false
+			],
+		];
+	}
+
+	/**
+	 * @dataProvider auditLogDataProvider
+	 * @param int $logLevel
+	 * @param string $pathToAuditLog
+	 * @param array<mixed> $logCondition
+	 * @param string $isAdminAuditAppInstalled
+	 * @param bool $expectedResult
+	 *
+	 * @return void
+	 */
+
+	public function testIsAdminAuditConfigSetCorrectly(int $logLevel, string $pathToAuditLog, array $logCondition, bool $isAdminAuditAppInstalled, bool $expectedResult): void {
+		$configMock = $this->getMockBuilder(IConfig::class)->getMock();
+		$iAppManagerMock = $this->getMockBuilder(IAppManager::class)->getMock();
+		$configMock
+			->method('getSystemValue')
+			->withConsecutive(
+				['loglevel'],
+				['logfile_audit'],
+				['log.condition']
+			)->willReturnOnConsecutiveCalls($logLevel, $pathToAuditLog, $logCondition);
+		$userManagerMock = $this->getMockBuilder(IUserManager::class)
+			->getMock();
+		$iAppManagerMock->method('isInstalled')->with('admin_audit')
+			->willReturn($isAdminAuditAppInstalled);
+		$service = $this->getServiceMock(
+			[],
+			null,
+			null,
+			$userManagerMock,
+			null,
+			$iAppManagerMock,
+			null,
+			null,
+			$configMock
+		);
+		$actualResult = $service->isAdminAuditConfigSetCorrectly();
+		$this->assertEquals($expectedResult, $actualResult);
 	}
 }